lib/pydantic-ai/requirements.txt
# PydanticAI Agent Server Requirements

# Core dependencies
pydantic-ai==0.0.14
pydantic==2.10.3
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Azure OpenAI
openai==1.57.4

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
